Android: Fix for 64bit arm and x86_64 targets
tested on kitkat arm and lollipop x86_64/arm64

Conflicts:
	Android.mk
	android/librpc/bindresvport.c

+++ android/libc/__set_errno.c
@@ -34,7 +34,7 @@
  * but is no more present for 64-bit targets!
  */
 
-int __set_errno(int n)
+long __set_errno(int n)
 {
     errno = n;
     return -1;
